requirements.txt
streamlit==1.42.0
trafilatura==2.0.0
spacy==3.8.3
numpy==1.26.4
faiss-cpu==1.9.0.post1
sentence_transformers==3.3.1
python-dotenv==1.0.1
groq==0.18.0